Got first benchmark working with Trenz board.
There was a bug in my status packet request. It was 1 word too
short, so the hardware wouldn't respond. I had thought that
the status request worked fine since it works with the dummy
hardware, so I made a number of changes in debugging the network
before I worked this out.